Online-Reviews
電話番号のカラムを’unsignedBigInteger'に変更。DBの許容を超えたためエラーが出ていた。

+++ database/migrations/2021_01_20_115743_add_admin_to_users_table.php
@@ -21,7 +21,7 @@ class AddAdminToUsersTable extends Migration
             $table->integer('birthday')->after('age');//誕生日
             $table->integer('postal_code')->after('birthday');//郵便番号
             $table->string('address')->after('postal_code');//住所
-            $table->integer('phone')->after('address');//電話番号
+            $table->unsignedBigInteger('phone')->after('address');//電話番号
         });
     }
